requirements.txt
langchain
ctransformers
flask
sentence-transformers
langchain-community
dotenv-python
pinecone
langchain-pinecone
langchain-community
langchain-huggingface
pypdf

-e .
